+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Released]
+### 0.0.4 - 2018-03-16
+#### Changed
+- fixed pdf output location variable
+
+#### Added
+- basic functions: {
+    - command for search pdf files and move then to destination
+}
+
 ### 0.0.3 - 2018-03-16
 #### Changed
 - added try/except to pisa pdf line